servers/verifier/pyproject.toml
[project]
name = "network-verifier"
version = "0.1.0"
description = "MCP server for network configuration verification"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]",
    "pybatfish",
    "pydantic",
    "pyyaml"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
